remove the IWCVV/IWCV distinction
