+++ shell/browser/api/electron_api_web_contents_view.cc
@@ -28,7 +28,7 @@ namespace electron::api {
 
 WebContentsView::WebContentsView(v8::Isolate* isolate,
                                  gin::Handle<WebContents> web_contents)
-    : View(web_contents->inspectable_web_contents()->GetView()->GetView()),
+    : View(web_contents->inspectable_web_contents()->GetView()),
       web_contents_(isolate, web_contents.ToV8()),
       api_web_contents_(web_contents.get()) {
   set_delete_view(false);
