+++ shell/browser/ui/inspectable_web_contents_view.h
@@ -9,27 +9,28 @@
 #include <string>
 
 #include "base/memory/raw_ptr.h"
+#include "chrome/browser/devtools/devtools_contents_resizing_strategy.h"
 #include "electron/shell/common/api/api.mojom.h"
 #include "ui/gfx/native_widget_types.h"
+#include "ui/views/view.h"
 
 class DevToolsContentsResizingStrategy;
-
-#if defined(TOOLKIT_VIEWS)
 namespace views {
-class View;
-}
-#endif
+class WebView;
+class Widget;
+class WidgetDelegate;
+}  // namespace views
 
 namespace electron {
 
 class InspectableWebContents;
 class InspectableWebContentsViewDelegate;
 
-class InspectableWebContentsView {
+class InspectableWebContentsView : public views::View {
  public:
   explicit InspectableWebContentsView(
       InspectableWebContents* inspectable_web_contents);
-  virtual ~InspectableWebContentsView();
+  ~InspectableWebContentsView() override;
 
   InspectableWebContents* inspectable_web_contents() {
     return inspectable_web_contents_;
@@ -41,19 +42,21 @@ class InspectableWebContentsView {
   }
   InspectableWebContentsViewDelegate* GetDelegate() const { return delegate_; }
 
-  // Returns the container control, which has devtools view attached.
-  virtual views::View* GetView() = 0;
+  void ShowDevTools(bool activate);
+  void CloseDevTools();
+  bool IsDevToolsViewShowing();
+  bool IsDevToolsViewFocused();
+  void SetIsDocked(bool docked, bool activate);
+  void SetContentsResizingStrategy(
+      const DevToolsContentsResizingStrategy& strategy);
+  void SetTitle(const std::u16string& title);
+  const std::u16string GetTitle();
 
-  virtual void ShowDevTools(bool activate) = 0;
-  // Hide the DevTools view.
-  virtual void CloseDevTools() = 0;
-  virtual bool IsDevToolsViewShowing() = 0;
-  virtual bool IsDevToolsViewFocused() = 0;
-  virtual void SetIsDocked(bool docked, bool activate) = 0;
-  virtual void SetContentsResizingStrategy(
-      const DevToolsContentsResizingStrategy& strategy) = 0;
-  virtual void SetTitle(const std::u16string& title) = 0;
-  virtual const std::u16string GetTitle() = 0;
+  // views::View:
+  void Layout() override;
+#if BUILDFLAG(IS_MAC)
+  bool OnMousePressed(const ui::MouseEvent& event) override;
+#endif
 
  protected:
   // Owns us.
@@ -62,6 +65,16 @@ class InspectableWebContentsView {
  private:
   raw_ptr<InspectableWebContentsViewDelegate> delegate_ =
       nullptr;  // weak references.
+
+  std::unique_ptr<views::Widget> devtools_window_;
+  raw_ptr<views::WebView> devtools_window_web_view_ = nullptr;
+  raw_ptr<views::View> contents_web_view_ = nullptr;
+  raw_ptr<views::WebView> devtools_web_view_ = nullptr;
+
+  DevToolsContentsResizingStrategy strategy_;
+  bool devtools_visible_ = false;
+  raw_ptr<views::WidgetDelegate> devtools_window_delegate_ = nullptr;
+  std::u16string title_;
 };
 
 }  // namespace electron
